Global Land-Ocean Temperature Prediction
The emission of large quantities of greenhouse gases, such as carbon dioxide, has led to global warming. during the twentieth century, the world's average temperature rose by about 0.6 degrees Celsius. The dangers of warming range from natural disasters to the rupture of biological chains, and touch on every aspect of human existence. In this report, the ARIMA model will be used to predict the trend of global temperature change in 2022-2024.